requirements.txt
discord.py>=2.3.2
pymongo>=4.6.1
motor>=3.3.2
aiohttp>=3.9.1
python-dotenv>=1.0.0
colorama>=0.4.6
typing-extensions>=4.9.0
dnspython>=2.4.2
python-dateutil>=2.8.2
pytz>=2023.3.post1
requests>=2.31.0
urllib3>=2.1.0
websockets>=12.0
yarl>=1.9.4